fix: cleaned up the code, moved functions in individual files
