+++ src/api/updateGateway.ts
@@ -6,9 +6,7 @@ import {
 import { fromUtf8, toUtf8 } from '@aws-sdk/util-utf8-browser'
 import type { Static } from '@sinclair/typebox'
 import type {
-	DesiredGameState,
 	DesiredRobot as DesiredRobotSchema,
-	ReportedGameState,
 	ReportedRobot as ReportedRobotSchema,
 } from 'api/validateGameControllerShadow'
 import { validateGameControllerShadow } from 'api/validateGameControllerShadow.js'
@@ -78,30 +76,3 @@ export const updateGatewayController =
 				console.error(error)
 			})
 	}
-
-export const calculateReported = ({
-	robots,
-	reported,
-}: {
-	reported: Static<typeof ReportedGameState>
-	robots: Required<Static<typeof DesiredGameState>>['robots']
-}): Static<typeof ReportedGameState> => {
-	const reportedGameState: Static<typeof ReportedGameState> = {
-		...reported,
-	}
-
-	for (const [robotMac, command] of Object.entries(robots)) {
-		const updatedRobot: ReportedRobot = {
-			...reported.robots[robotMac],
-			angleDeg: command.angleDeg,
-			driveTimeMs: command.driveTimeMs,
-			revolutionCount: 0,
-		}
-		reportedGameState.robots = {
-			...reportedGameState.robots,
-			[robotMac]: updatedRobot,
-		}
-	}
-
-	return reportedGameState
-}
