+++ src/utils/initialReportedState.ts
@@ -1,0 +1,45 @@
+export const robotCommands = {
+	round: 1,
+	robots: {
+		'00:25:96:FF:FE:12:34:51': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:52': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:53': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:54': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:55': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:56': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:57': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+		'00:25:96:FF:FE:12:34:58': {
+			angleDeg: 0,
+			driveTimeMs: 0,
+			revolutionCount: 0,
+		},
+	},
+}
